Read passwords from file if file exists
